Add slowness warning

+++ include/localization/constants.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-constexpr double MM_TO_IN = 1 / 25.4;
-constexpr double M_TO_IN = 39.3701;
+constexpr float MM_TO_IN = 1 / 25.4;
+constexpr float M_TO_IN = 39.3701;
 
 // Value from alexDickhans/echo on GitHub
 constexpr float WALL = 1.78308 * M_TO_IN;
